+++ std-bits/table/src/main/java/org/enso/table/data/column/builder/TimeOfDayBuilder.java
@@ -21,7 +21,7 @@ final class TimeOfDayBuilder extends TypedBuilder<LocalTime> {
     var bits = BitSet.valueOf(validityBuffer);
     var buf =
         MemorySegment.ofAddress(data)
-            .reinterpret((long) Long.BYTES * size)
+            .reinterpret(Long.BYTES * size)
             .asByteBuffer()
             .order(ByteOrder.LITTLE_ENDIAN);
